+++ tests/Transformers/KNNImputerTest.php
@@ -1,0 +1,73 @@
+<?php
+
+namespace Rubix\ML\Tests\Transformers;
+
+use Rubix\ML\Datasets\Unlabeled;
+use Rubix\ML\Transformers\Elastic;
+use Rubix\ML\Transformers\Stateful;
+use Rubix\ML\Transformers\KNNImputer;
+use Rubix\ML\Transformers\Transformer;
+use Rubix\ML\Datasets\Generators\Blob;
+use Rubix\ML\Kernels\Distance\SafeEuclidean;
+use PHPUnit\Framework\TestCase;
+use RuntimeException;
+
+class KNNImputerTest extends TestCase
+{
+    protected const RANDOM_SEED = 0;
+
+    protected $dataset;
+
+    protected $generator;
+
+    protected $transformer;
+
+    public function setUp()
+    {
+        $this->dataset = new Unlabeled([
+            [30, 0.001],
+            [NAN, 0.055],
+            [50, -2.0],
+            [60, NAN],
+            [10, 1.0],
+            [100, 9.0],
+        ]);
+
+        $this->generator = new Blob([30., 0.]);
+
+        $this->transformer = new KNNImputer(5, true, new SafeEuclidean());
+
+        srand(self::RANDOM_SEED);
+    }
+
+    public function test_build_transformer()
+    {
+        $this->assertInstanceOf(KNNImputer::class, $this->transformer);
+        $this->assertInstanceOf(Transformer::class, $this->transformer);
+        $this->assertInstanceOf(Stateful::class, $this->transformer);
+        $this->assertInstanceOf(Elastic::class, $this->transformer);
+    }
+
+    public function test_fit_update_transform()
+    {
+        $this->transformer->fit($this->dataset);
+
+        $this->assertTrue($this->transformer->fitted());
+
+        $this->transformer->update($this->generator->generate(30));
+
+        $this->dataset->apply($this->transformer);
+
+        $this->assertEquals(30.1161413983789, $this->dataset[1][0]);
+        $this->assertEquals(-0.6514400548002053, $this->dataset[3][1]);
+    }
+
+    public function test_transform_unfitted()
+    {
+        $this->expectException(RuntimeException::class);
+
+        $samples = $this->dataset->samples();
+
+        $this->transformer->transform($samples);
+    }
+}